requirements.txt
pytest>=7.4.0
pytest-html>=4.0.0
pytest-cov>=4.1.0
pyyaml>=6.0.1
pandas>=2.0.0
matplotlib>=3.7.0
cantools>=39.0.0
python-can>=4.3.0
numpy>=1.24.0
jinja2>=3.1.0
